made changes to file structure and added mock token
