+++ frontend/src/App/index.js
@@ -1,11 +1,1 @@
-import React, { Component } from 'react';
-import Routes from '../Routes';
-import { withRouter } from 'react-router-dom';
-
-class App extends Component {
-  render() {
-    return <Routes />;
-  }
-}
-
-export default App;
+export {default} from './App';
