+++ frontend/src/Button/index.js
@@ -1,14 +1,1 @@
-import React, { Component } from 'react';
-import styled from 'styled-components';
-
-class Button extends Component {
-  render() {
-    return <StyledButton />;
-  }
-}
-
-const StyledButton = styled.button`
-  color: tomato;
-`;
-
-export default Button;
+export {default} from './Button';
