+++ frontend/src/Checkin/CheckIn.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class CheckIn extends Component {
+  render() {
+    return <div>ok</div>;
+  }
+}
+
+export default CheckIn;
