+++ frontend/src/Checkin/index.js
@@ -1,9 +1,1 @@
-import React, { Component } from 'react';
-
-class Checkin extends Component {
-  render() {
-    return <div>ok</div>;
-  }
-}
-
-export default Checkin;
+export {default} from './CheckIn';
