+++ frontend/src/PlayerList/index.js
@@ -1,0 +1,2 @@
+export {default} from './PlayerList';
+
